Move findOperators into shim layer due to changes in ShuffleExchangeExec between spark versions
