+++ tests/src/test/scala/com/nvidia/spark/rapids/BroadcastNestedLoopJoinSuite.scala
@@ -35,7 +35,7 @@ class BroadcastNestedLoopJoinSuite extends SparkQueryCompareTestSuite {
       df3.collect()
       val plan = df3.queryExecution.executedPlan
 
-      val nljCount = GpuTransitionOverrides
+      val nljCount = ShimLoader.getSparkShims
         .findOperators(plan, _.isInstanceOf[GpuBroadcastNestedLoopJoinExecBase])
       assert(nljCount.size === 1)
     }, conf)
@@ -53,7 +53,7 @@ class BroadcastNestedLoopJoinSuite extends SparkQueryCompareTestSuite {
       df3.collect()
       val plan = df3.queryExecution.executedPlan
 
-      val nljCount = GpuTransitionOverrides
+      val nljCount = ShimLoader.getSparkShims
         .findOperators(plan, _.isInstanceOf[GpuBroadcastNestedLoopJoinExecBase])
 
       ShimLoader.getSparkShims.getSparkShimVersion match {
